move everything around to try and make it cleaner
