+++ tests/fixtures/test_robot/test_tags_with_kwargs/conftest.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from pytest_robotframework._internal.pytest_robot_items import RobotItem
+    from pytest_robotframework._internal.pytest.robot_file_support import RobotItem
 
 
 def pytest_runtest_setup(item: RobotItem):
